tests/wpt/web-platform-tests/web-animations/animation/playState.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animation.playState</title>
<link rel="help" href="https://w3c.github.io/web-animations/#dom-animation-playstate">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
'use strict';

promise_test(function(t) {
  var div = createDiv(t);
  var animation = div.animate({}, 100 * MS_PER_SEC);

  assert_equals(animation.playState, 'pending');
  return animation.ready.then(function() {
    assert_equals(animation.playState, 'running');
  });
}, 'Animation.playState reports \'pending\'->\'running\' when initially ' +
   'played');

promise_test(function(t) {
  var div = createDiv(t);
  var animation = div.animate({}, 100 * MS_PER_SEC);
  animation.pause();

  assert_equals(animation.playState, 'pending');
  return animation.ready.then(function() {
    assert_equals(animation.playState, 'paused');
  });
}, 'Animation.playState reports \'pending\'->\'paused\' when pausing');

test(function(t) {
  var div = createDiv(t);
  var animation = div.animate({}, 100 * MS_PER_SEC);
  animation.cancel();
  assert_equals(animation.playState, 'idle');
}, 'Animation.playState is \'idle\' when canceled.');

test(function(t) {
  var div = createDiv(t);
  var animation = div.animate({}, 100 * MS_PER_SEC);
  animation.cancel();
  animation.currentTime = 50 * MS_PER_SEC;
  assert_equals(animation.playState, 'paused',
                'After seeking an idle animation, it is effectively paused');
}, 'Animation.playState is \'paused\' after cancelling an animation, ' +
   'seeking it makes it paused');

</script>
</body>
